requirements.txt
Pillow
PyYAML
pymediainfo==6.0.1
pylrc==0.1.2
